Make restore tool backward compatiable
In commit 4e5cbcc8574b8258c0146d4d01539b6d48212f4a, we have changed
the way we write backup to file. This also changes restore process.
This was breaking change as we could not restore, old backups using
the code from this commit onwards. Current commit adds backward
compatibility to Badger. While unmarshalling a buffer, it will try
to unmarshal first as list. If there is err in unmarshalling, it will
try to unmarshal it as single KV.